RES-2502: Move self supervised learning framework to standalone package
