+++ nupic/research/frameworks/vernon/distributed/experiments/__init__.py
@@ -22,4 +22,3 @@
 from .cl_experiment import *
 from .meta_cl_experiment import *
 from .supervised_experiment import *
-from .self_supervised_experiment import *
