+++ nupic/research/frameworks/vernon/experiments/__init__.py
@@ -21,5 +21,4 @@
 
 from .cl_experiment import *
 from .meta_cl_experiment import *
-from .self_supervised_experiment import *
 from .supervised_experiment import *
